Exercise 4.12 blog list expansion Step 9: Change the delete operation so that a blog can only be deleted by the user that added it.
(Also fixed testing suite for other operations and added tests for delete)

+++ controllers/blogs.js
@@ -35,8 +35,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'Invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'Cannot delete blogs created by another user' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
